requirements.txt
APScheduler==3.6.3
astroid==2.5.6
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
idna==2.10
imageio==2.9.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
numpy==1.20.3
Pillow==8.2.0
pylint==2.8.2
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.3
tzlocal==2.1
urllib3==1.26.5
wrapt==1.12.1
